>CesGridViewPro 1. Enable/Disable visibility of filtering row.
>CesTextBox
1. Change return type to decimal for number type input.

>CesListBox
1. "RightToLeft" property implemented

>CesComboBox
1. "RightToLeft" property implemented

+++ Ces.WinForm.UI/CesTextBox.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel;
-using System.Drawing;
 
 namespace Ces.WinForm.UI
 {
@@ -384,19 +383,18 @@ namespace Ces.WinForm.UI
             Text = string.Empty;
         }
 
-        public double GetValue()
+        public decimal GetValue()
         {
             if(string.IsNullOrEmpty(CesText?.Trim()))
                 return 0;
 
             var normalizedText = CesText.Trim().Replace(",", string.Empty);
-
-            var check = double.TryParse(normalizedText, out double result);
+            var check = decimal.TryParse(normalizedText, out decimal result);
 
             if (!check)
                 return result;
             else
-                return double.Parse(normalizedText);
+                return decimal.Parse(normalizedText);
         }
 
         private void SetTextBoxWidth()
